requirements.txt
Django
djangorestframework
notebook
pandas
